// Contacts/api/generated/api.ts
export interface Address {
    /**
     * 
     * @type {number}
     * @memberof Address
     */
    addressId?: number;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    streetAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    secondaryAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    unit?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    city?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    state?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    country?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Address
     */
    postalCode?: string | null;
    /**
     * 
     * @type {AddressType}
     * @memberof Address
     */
    addressType?: AddressType;
}
/**
 * 
 * @export
 * @interface AddressType
 */
export interface AddressType {
    /**
     * 
     * @type {number}
     * @memberof AddressType
     */
    addressTypeId?: number;
    /**
     * 
     * @type {string}
     * @memberof AddressType
     */
    type?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressType
     */
    description?: string | null;
}
/**
 * 
 * @export
 * @interface Contact
 */
export interface Contact {
    /**
     * 
     * @type {number}
     * @memberof Contact
     */
    contactId?: number;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    middleName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    emailAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    birthday: string;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    anniversary?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {Array<Address>}
     * @memberof Contact
     */
    addresses?: Array<Address> | null;
    /**
     * 
     * @type {Array<Phone>}
     * @memberof Contact
     */
    phones?: Array<Phone> | null;
    /**
     * 
     * @type {string}
     * @memberof Contact
     */
    fullName?: string | null;
}
/**
 * 
 * @export
 * @interface Phone
 */
export interface Phone {
    /**
     * 
     * @type {number}
     * @memberof Phone
     */
    phoneId?: number;
    /**
     * 
     * @type {string}
     * @memberof Phone
     */
    phoneNumber?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Phone
     */
    extension?: string | null;
    /**
     * 
     * @type {PhoneType}
     * @memberof Phone
     */
    phoneType?: PhoneType;
}
/**
 * 
 * @export
 * @interface PhoneType
 */
export interface PhoneType {
    /**
     * 
     * @type {number}
     * @memberof PhoneType
     */
    phoneTypeId?: number;
    /**
     * 
     * @type {string}
     * @memberof PhoneType
     */
    type?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PhoneType
     */
    description?: string | null;
}